mymine/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-graphql-jwt==0.4.0
celery==5.4.0
django-channels==0.7.0
django-simple-history==3.7.0
sqlalchemy==2.0.36
redis==5.2.1
pytest==8.3.4
coverage==7.6.9
graphene-django==3.2.2
python-dateutil==2.9.0
requests==2.32.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.1
